+++ src/time/instant.rs
@@ -50,6 +50,10 @@ impl Instant {
 			TIME_ORIGIN.with(|origin| performance.now() + origin)
 		});
 
+		assert!(
+			now.is_sign_positive(),
+			"negative `DOMHighResTimeStamp`s are not supported"
+		);
 		Self(time_stamp_to_duration(now))
 	}
 
@@ -179,7 +183,7 @@ fn time_stamp_to_duration(time_stamp: f64) -> Duration {
 	let time_stamp = F64(time_stamp);
 
 	Duration::from_millis(time_stamp.trunc() as u64)
-		+ Duration::from_nanos(F64(time_stamp.fract() * 1.0e6).round() as u64)
+		+ Duration::from_nanos(F64(time_stamp.fract() * 1.0e6).internal_round_ties_even() as u64)
 }
 
 /// [`f64`] `no_std` compatibility wrapper.
@@ -235,32 +239,53 @@ impl F64 {
 		self.0 - self.trunc()
 	}
 
-	/// See [`f64::copysign()`].
-	///
-	/// [`f64::copysign()`]: https://doc.rust-lang.org/std/primitive.f64.html#method.copysign
-	#[cfg(not(feature = "std"))]
-	fn copysign(self, sign: f64) -> f64 {
-		// See <https://github.com/rust-lang/libm/blob/libm-v0.2.11/src/math/copysign.rs>.
-
-		let mut ux = self.0.to_bits();
-		let uy = sign.to_bits();
-		ux &= (!0) >> 1;
-		ux |= uy & (1 << 63);
-		f64::from_bits(ux)
-	}
-
-	/// See [`f64::round()`].
-	#[cfg(feature = "std")]
+	/// See [`f64::round_ties_even()`].
 	#[allow(clippy::disallowed_methods)]
-	fn round(self) -> f64 {
-		self.0.round()
+	#[cfg(all(feature = "std", feature = "msrv"))]
+	#[rustversion::since(1.77)]
+	fn internal_round_ties_even(self) -> f64 {
+		#[allow(clippy::incompatible_msrv)]
+		self.0.round_ties_even()
 	}
 
-	#[cfg(not(feature = "std"))]
-	fn round(self) -> f64 {
-		// See <https://github.com/rust-lang/libm/blob/libm-v0.2.11/src/math/round.rs>.
+	/// A specialized version of [`f64::round_ties_even()`]. [`f64`] must be
+	/// positive and have an exponent smaller than `52`.
+	///
+	/// - We expect `DOMHighResTimeStamp` to always be positive. We check that
+	///   in [`Instant::now()`].
+	/// - We only round the fractional part after multiplying it by `1e6`. A
+	///   fraction always has a negative exponent. `1e6` has an exponent of
+	///   `19`. Therefor the resulting exponent can at most be `19`.
+	///
+	/// [`f64::round_ties_even()`]: https://doc.rust-lang.org/1.83.0/std/primitive.f64.html#method.round_ties_even
+	#[cfg_attr(all(feature = "std", feature = "msrv"), rustversion::before(1.77))]
+	fn internal_round_ties_even(self) -> f64 {
+		/// Put `debug_assert!` in a function to clap `coverage(off)` on it.
+		///
+		/// See <https://github.com/rust-lang/rust/issues/80549>.
+		#[cfg_attr(wasm_bindgen_unstable_test_coverage, coverage(off))]
+		fn check(this: f64) {
+			debug_assert!(this.is_sign_positive(), "found negative input");
+			debug_assert!(
+				{
+					let exponent: u64 = this.to_bits() >> 52 & 0x7ff;
+					exponent < 0x3ff + 52
+				},
+				"found number with exponent bigger than 51"
+			);
+		}
 
-		Self(self.0 + Self(0.5 - 0.25 * f64::EPSILON).copysign(self.0)).trunc()
+		check(self.0);
+
+		// See <https://github.com/rust-lang/libm/blob/libm-v0.2.11/src/math/rint.rs>.
+
+		let one_over_e = 1.0 / f64::EPSILON;
+		// REMOVED: We don't support numbers with exponents bigger than 51.
+		// REMOVED: We don't support negative numbers.
+		// REMOVED: We don't support numbers with exponents bigger than 51.
+		let xplusoneovere = self.0 + one_over_e;
+		xplusoneovere - one_over_e
+		// REMOVED: We don't support negative numbers.
 	}
 }